Rename initialise module entrypoint
