+++ src/brightcove_async/__init__.py
@@ -1,5 +1,5 @@
 from . import schemas
 from .client import BrightcoveClient
-from .initalise import initialise_brightcove_client
+from .initialise import initialise_brightcove_client
 
 __all__ = ["BrightcoveClient", "initialise_brightcove_client", "schemas"]
